SHORTY-87 Fix 500 http code to 400 in case SignUp case with invalid age

+++ src/main/kotlin/com/zufar/urlshortener/auth/controller/AuthController.kt
@@ -268,7 +268,7 @@ class AuthController(
             email = signUpRequest.email,
             password = passwordEncoder.encode(signUpRequest.password),
             country = signUpRequest.country,
-            age = signUpRequest.age,
+            age = signUpRequest.age.toInt(),
             createdAt = LocalDateTime.now(),
             updatedAt = LocalDateTime.now()
         )
